add admin claim

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -44,6 +44,9 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name): 
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"error_message": "Only Admin can delete data"}, 401 
         item = ItemModel.find_by_name(name)
         if item: 
             item.delete_from_db()
